requirements.txt
annotated-types==0.7.0
-e git+ssh://git@github.com/an3317/BIOADAPT.git@d830f610f5d675c0f63aeeab195efc650f3720c8#egg=bioadapt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
earthengine-api==1.5.12
fonttools==4.57.0
fpdf2==2.7.8
geeutils==2.0.0
geojson==3.2.0
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.1.0
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.9.0
multimethod==1.10
mypy_extensions==1.1.0
numba==0.61.2
numpy==1.26.4
packaging==25.0
pandas==2.2.2
pandera==0.18.3
pillow==11.2.1
pipeline==0.1.0
pipeline-utils==0.0.2
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
scikit-learn==1.5.0
scipy==1.15.2
seaborn==0.13.2
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.66.4
typeguard==4.4.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
wrapt==1.17.2
xgboost==2.0.3
